Move expandLastSelectedRegion back to selectable.tsx
