+++ packages/table/src/common/internal/selectionUtils.ts
@@ -88,33 +88,3 @@ export function resizeRegion(region: IRegion, direction: Direction, focusedCell?
     // sanitizing the result.
     return nextRegion;
 }
-
-/**
- * Expands the last-selected region to the new region, and replaces the
- * last-selected region with the expanded region. If a focused cell is provided,
- * the focused cell will serve as an anchor for the expansion.
- */
-export function expandLastSelectedRegion(
-    selectedRegions: IRegion[],
-    newRegion: IRegion,
-    focusedCell?: IFocusedCellCoordinates,
-) {
-    if (selectedRegions.length === 0) {
-        return [newRegion];
-    } else if (focusedCell != null) {
-        const expandedRegion = FocusedCellUtils.expandFocusedRegion(focusedCell, newRegion);
-        return Regions.update(selectedRegions, expandedRegion);
-    } else {
-        const expandedRegion = Regions.expandRegion(selectedRegions[selectedRegions.length - 1], newRegion);
-        return Regions.update(selectedRegions, expandedRegion);
-    }
-}
-
-/**
- * Returns the last region in the provided `selectedRegions` array, or `undefined`.
- */
-export function getLastSelectedRegion(selectedRegions: IRegion[]) {
-    return selectedRegions == null || selectedRegions.length === 0
-        ? undefined
-        : selectedRegions[selectedRegions.length - 1];
-}
